Create Guessing a number.py
The Python Number Guessing Game is a basic program where the player has to make guesses for a different number selected in advance. The feedback from the program is given on each of the guesses - too high or too low - thus assisting the player to eliminate options. In this first version, the word is with simple objective - the player has to guess the correct number and the computer congratulates the player. It is a fun way of practicing logical thinking and could be enriched with such options as guess limits or varying difficulty, monitoring the number of attempts, etc.